Analyzing Genetic Deviations
The rapid growth and massive volume of human genomic data, especially Single Nucleotide Polymorphisms (SNPs), presents challenges for biomedical researchers as well as programmers of machine learning algorithms. One such challenge is to identify an optimal subset of SNPs, to find the most haplotype diversity for each haplotype block. This information minimization process facilitates cost-effective genotyping, and thus the study of genotype-phenotype association. It also has implications for assessing the risk of identifying research subjects on the basis of SNP information available in the electronic gene bank.
In this paper, a method is proposed to select features for SNPs by adopting a new kernel in the KPCA "Primary Component Analysis Algorithm" for the purpose of discrimination, as well as combining it with the "Particle Swarm Optimization Intelligent Search Algorithm PSO", which is named PSO_KPCA. The proposed method was applied to five real experimental data from the global HapMap website, and it was found that the proposed method achieved an accuracy of 98.22% in case the affected SNPs were distinguished. The proposed method was presented in an integrated software model available on the GUI and written in Matlab®2016a, that reads the data file and has the extension *.hmp, and a number of comprehensive exploratory analyzes of the data are performed at the genome level.